requirements.txt
selenium
pandas
pyvirtualdisplay
psycopg2-binary
